+++ internal/provider/log_export_config_resource.go
@@ -171,7 +171,7 @@ func (r *logExportConfigResource) Create(
 		return
 	}
 
-	if cluster.Config.Serverless != nil {
+	if cluster.Config.Shared != nil {
 		resp.Diagnostics.AddError(
 			"Incompatible cluster type",
 			"Log export services are only available for dedicated clusters",